Add Docker environment files and configuration for metricbeat and filebeat
A docker-compose.yml file was modified to add two new environment files, `.env` and `.env.local`, for the metricbeat and filebeat services. These environment files will be used to provide additional configuration to the services.

In addition, a new file called `filebeat.yml` was added. This file contains the configuration for filebeat, including file